requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
exceptiongroup==1.2.2
fonttools==4.57.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.3.2
numpy==2.2.4
openai==0.28.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
